Reticle Improvements / Level 2 Building
Made the crosshair slowly turn red when targeting an enemy and slowly turn blue when targeting a civilian

Built a full outline for level 2 and added the player to it; some things still are not fully functional with the player in level 2 but you can move around. Also added textures to walls and floors.

Added some props to Level 2, including outside, bathroom, and sign.

Created a few prefabs for building pieces to make the transition from Level 1 to building Level 2 easier.