Adding Main Menu and Pause Menu + Functionality
Added Main Menu with five options. Can change sensitivity in the settings menu; the SettingsManager holds the value which is taken by the PlayerMovement Script. Sensitivity is saved as a player preference. Can check info about the game devs. Can also see how long you have been playing the game in minutes; this is stored with the TimeManager script. The pause menu works like how we did in class; I updated a couple other scripts so that they would functionally pause when needed and unpause when needed. There is a bug with pausing where sometimes the player is frozen at the beginning of the game.